+++ src/Learning/KWData/CMakeLists.txt
@@ -5,8 +5,8 @@
 if(BUILD_LEX_YACC)
   bison_target(KWCParser KWCYac.yac ${CMAKE_CURRENT_SOURCE_DIR}/KWCYac.cpp)
   if(MSVC)
-    flex_target(KWCScanner KWCLex.lex KWCLex.inc COMPILE_FLAGS-- nounistd)
-    flex_target(JsonScanner JsonLex.lex JsonLex.inc COMPILE_FLAGS-- nounistd)
+    flex_target(KWCScanner KWCLex.lex KWCLex.inc COMPILE_FLAGS "--nounistd")
+    flex_target(JsonScanner JsonLex.lex JsonLex.inc COMPILE_FLAGS "--nounistd")
   else()
     flex_target(KWCScanner KWCLex.lex KWCLex.inc)
     flex_target(JsonScanner JsonLex.lex JsonLex.inc)
